requirements.txt
docker>=5.0.0
flask>=2.0.0
pyyaml>=6.0
werkzeug>=2.0.0
anthropic>=0.8.1
python-dotenv>=1.0.0 